Remove split between libdepthmapx and the Gui App - this was only ever introduced for testing and is causing issues with cmake
