+++ depthmapX/coreapplication.cpp
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "GuiApp/coreapplication.h"
+#include "depthmapX/coreapplication.h"
 #include <QDesktopWidget>
 
 int CoreApplication::exec() {
