+++ depthmapX/coreapplication.h
