+++ depthmapX/resources/Info.plist
